🐛 FIx: notice페이지 모듈 오류 수정
