+++ src/pages/notice/notice.module.css
@@ -1,0 +1,181 @@
+/* notice2.module.css */
+.content {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  overflow-y: auto;
+}
+
+.title {
+  margin-left: 60px;
+  font-size: 35px;
+  font-weight: 500;
+}
+
+.cards {
+  display: flex;
+  grid-template-columns: repeat(auto-fill, minmax(200px, calc((100% - 4.5rem) / 4)));
+  gap: 2rem;
+  margin: 0 0.5rem;
+  overflow-y: auto;
+  justify-content: center;
+  flex-wrap: wrap;
+}
+
+.card {
+  width: 100%;
+  min-width: 200px;
+  max-width: 330px;
+  height: 600px;
+  background-color: #055D20;
+  border-radius: 20px;
+  text-align: center;
+  cursor: pointer;
+  padding: 0.8rem;
+  transition: transform 0.3s ease;
+  aspect-ratio: 3/5;
+  margin: 1rem 0;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
+}
+
+.card:hover {
+  transform: scale(1.05);
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.12);
+}
+
+.card h2 {
+  font-size: 1.3rem;
+  font-weight: 500;
+  color: #f8bd64;
+}
+
+.card p {
+  margin: 0.5rem 0;
+  line-height: 1.6;
+  color: #ffffff;
+  font-size: 1rem;
+}
+
+.imagePlaceholder {
+  width: 100%;
+  height: 450px;
+  object-fit: cover;
+  border-radius: 15px;
+  margin-bottom: 10px;
+  background-color: #828282;
+}
+
+.cards::-webkit-scrollbar,
+.modalTextContent::-webkit-scrollbar {
+  width: 8px;
+}
+
+.cards::-webkit-scrollbar-thumb,
+.modalTextContent::-webkit-scrollbar-thumb {
+  background-color: #b1b1b1;
+  border-radius: 10px;
+}
+
+.cards::-webkit-scrollbar-track,
+.modalTextContent::-webkit-scrollbar-track {
+  background-color: #ffffff;
+}
+
+.modalContainer {
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.5);
+  z-index: 1;
+  transition: background-color 0.3s ease;
+  padding: 2rem;
+}
+
+.modalContent {
+  position: relative;
+  background-color: #fff;
+  width: 100%;
+  max-width: 1000px;
+  aspect-ratio: 2/5;
+  max-height: 70vh;
+  padding: 2.5rem 1rem 1rem;
+  border-radius: 2em;
+  display: flex;
+  flex-direction: column;
+  transform: translateY(20px);
+  transition: all 0.3s ease;
+  overflow: hidden;
+  background-color: #fefbf0;
+}
+
+.modalArea {
+  flex: 1;
+  padding: 1rem;
+  display: flex;
+  gap: 20px;
+  align-items: center;
+  overflow: hidden;
+  height: 100%;
+}
+
+.modalImage {
+  flex: 0 0 45%;
+  width: 280px;
+  height: 100%;
+  margin-top: 0.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.modalImage img {
+  width: 100%;
+  height: auto;
+  display: block;
+  border-radius: 1em;
+  object-fit: contain;
+}
+
+.modalTextContent {
+  flex: 1;
+  padding: 0 10px;
+  display: flex;
+  flex-direction: column;
+  overflow-y: auto;
+  height: 100%;
+}
+
+.modalTextContent h2 {
+  font-size: 1.8rem;
+  margin: 0.5rem 0;
+  font-weight: 500;
+}
+
+.modalTextContent p {
+  font-size: 1.3rem;
+  white-space: pre-line;
+  color: #555555;
+}
+
+.modalCloseButton {
+  position: absolute;
+  top: 20px;
+  right: 20px;
+  cursor: pointer;
+  font-size: 2rem;
+  color: #055D20;
+  transition: color 0.3s ease;
+}
+
+.modalCloseButton:hover {
+  color: #033412;
+}
+
+.hidden {
+  display: none;
+}
